Move library files into lib directory
This prevents them from showing up as executables when putting wit on
PATH
Also fix up some imports
